Method Overloading
we use the same name for the method, but we change for example the quantity of variables, or parameters.
method overloading increases the readability of the program, 
How can we achieve the method overloading? -->> by changing the number of arguments and the other option is by changing the data types.
see file Addition as an example.